Diplom/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.5
defusedxml==0.7.1
Django==5.2.11
django-allauth==65.14.1
django-baton==5.1.1
django-cachalot==2.9.0
django-celery-beat==2.8.1
django-filter==25.2
django-jet-reboot==1.3.10
django-redis==6.0.0
django-rest==0.8.7
django-rest-passwordreset==1.5.0
django-silk==5.4.3
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
easy-thumbnails==2.10.1
gprof2dot==2025.4.14
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
Markdown==3.10.2
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.3
redis==7.1.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
sentry-sdk==2.52.0
setuptools==80.10.2
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
